Create app directory for launch application
