+++ cmd/main.go
@@ -1,48 +1,14 @@
 package main
 
 import (
-	"net/http"
-	"os"
-
-	"github.com/Nikby53/image-converter/internal/handler"
-	"github.com/Nikby53/image-converter/internal/repository"
-	"github.com/Nikby53/image-converter/internal/service"
-	"github.com/joho/godotenv"
+	"github.com/Nikby53/image-converter/internal/app"
 	_ "github.com/lib/pq"
 	"github.com/sirupsen/logrus"
-	"github.com/spf13/viper"
 )
 
 func main() {
-	logrus.SetFormatter(new(logrus.JSONFormatter))
-	if err := initConfig(); err != nil {
-		logrus.Fatalf("error initializing configs: %s", err.Error())
-	}
-	if err := godotenv.Load(); err != nil {
-		logrus.Fatalf("error loading env variables: %s", err.Error())
-	}
-	db, err := repository.NewPostgresDB(repository.Config{
-		Host:     viper.GetString("db.host"),
-		Port:     viper.GetString("db.port"),
-		Username: viper.GetString("db.username"),
-		DBName:   viper.GetString("db.dbname"),
-		SSLMode:  viper.GetString("db.sslmode"),
-		Password: os.Getenv("DB_PASSWORD"),
-	})
+	err := app.Run()
 	if err != nil {
-		logrus.Fatalf("failed to initialize db: %s", err.Error())
+		logrus.Fatal("failed to start app: ", err)
 	}
-	repos := repository.New(db)
-	services := service.New(repos)
-	srv := handler.NewServer(services)
-
-	if err := http.ListenAndServe(":8000", srv); err != nil && err != http.ErrServerClosed {
-		logrus.Printf("ListenAndServe(): %s", err)
-	}
-}
-
-func initConfig() error {
-	viper.AddConfigPath("internal/configs")
-	viper.SetConfigName("config")
-	return viper.ReadInConfig()
 }
