+++ internal/configs/config.yml
@@ -1,8 +1,8 @@
-port: "8000"
+port: ":8000"
 
 db:
   username: "postgres"
-  host: "172.17.0.2"
+  host: "127.0.0.1"
   port: "5432"
   dbname: "mytestdb"
   sslmode: "disable"
